Graft clouds.yaml correctly even if we need to insert in the middle.
We were using APPEND to attach project_id and user and password,
assuming the auth section is at the end of the clouds entry. There
is no guarantee for that to be the case though.
Make the injection more robust, by explicitly allowing to specify a
subentry auth and inject it there.

Signed-off-by: Kurt Garloff <kurt@garloff.de>